DL on Flink PyTorch support FlinkML API

+++ dl-on-flink-pytorch/src/test/resources/with_input.py
@@ -30,7 +30,7 @@ def main(context: Context):
 
     dataloader = DataLoader(pytorch_context.get_dataset_from_flink())
     for r, in dataloader:
-        output_tensors = [torch.zeros([1, 1], dtype=torch.long) for _ in
+        output_tensors = [torch.zeros([1, 1], dtype=torch.int32) for _ in
                          range(dist.get_world_size())]
         dist.all_gather(output_tensors, r)
         print(f"Rank {pytorch_context.get_rank()}: {output_tensors}")
